requirements.txt
flask
scikit-learn
pandas
azure-storage-blob
